seq: fixed the 'lost events bug' in safe mode
The fix is to declare efTest and efTestAndClear to be synchronization
points in safe mode. Both now check whether any of the variables sync'ed
with the event flag have been changed (have their dirty flag set) and
update the state set local copy of these variables. This happens
regardless of whether the call is part of a when() condition or inside
an action block.