snc: added new switch -W to turn on extra warnings
This is in preparation of supporting indirect calls which cause foreign
functions to be reported as undefined variables. Without -W (which is
the default) *no* "undefined variable" warnings will be given. This is
ok, since if they really are undefined the C compiler will flag them as
errors anyway.

This makes foreign declarations obsolete. With hindsight introducing
them was a mistake. I should probably deprecate them.