bugfix: do not call seq_monitor in pvAssign but reset dbch->monid [amended for branch-2-2]
The problem was a call seq_monitor(ch, FALSE) inside pvAssign after we
destroy the channel (if one exists). This will always fail (strange I
never noticed that) and as a result dbch->monid would not be reset to
NULL. That in turn means when the channel is re-assigned and connects,
seq_monitor(ch, TRUE) will not do anything because it thinks there is
already a valid subscription.

The fix is to remove the call and instead only reset dbch->monid to NULL.