snc: redesign compilation of pv init clauses and event mask generation
This is a major redesign patch that affects many parts of the compiler.

We now support a more general way of embedding channels (i.e. variables of
pv type) anywhere inside arbitrary combinations of arrays and structs. This
was already possible for the plain assign part (without giving a channel
name) because of the pv type constructor. But the traditional assign,
monitor, sync, and syncq clauses still supported only one-dimensional
arrays.

The the sort of expressions that are allowed in these clauses is generalized
to arbitrary combinations of array subscripts and struct member selections.
As it turns out, this is exactly the kind of expression we want to partially
evaluate in when-conditions in order to generate an approximation of the
minimal state event masks.

At the same time we now support a new pv initializer syntax that allows to
specify all the assign, monitor etc stuff in one place.

A new tree-like data structure named channel_node keeps track of exactly
where inside a variable channels (and event flags) are buried, while the
data structures for channels and event flags now contain the expression
leading from the variable down to them.