feature: state set local monitors for global channels
A monitor clause for a global channel inside a state set is now allowed and
causes the variable (or rather: memory location) to be updated only for the
state set.

This patch also removes assign, monitor, sync, and syncq clauses inside
states from the syntax. This was never very useful since the effect was
extended to the whole state set anyway. We remove it so we can assume that a
monitor is either global or state set local.