snc: allow mixed pv and non-pv initialiser aggregates
This was already allowed by the syntax, but the compiler crashed when it
actually encountered such code.

In order to implement this, we had to split (normal, non-pv) aggregate
initialisers for SNL arrays and structs into separate init actions for the
elements and members, respectively. This is because SNL structs with members
of pv type contain additional CH_ID members in the generated code, and these
get initialised in a separate pass before the value initialisations.

Incidentally, this also allows to generate correct code for event flag
initialisers (which were previously allowed by the syntax, too, but
ignored when generating code).

Note that monitor, sync, and syncq cannot yet be specified in pv
initialisers, only PV names, despite what an earlier patch comment
erroneously said. This feature will be implemented in a separate patch; it
also requires syntax changes.

Also note that this patch changes the semantics of initialisers for SNL
structs that contain foreign (escaped C code) members: such members are
completely ignored when generating code for an aggregate initialiser for the
struct. Valid SNL code must initialise them by other means e.g. assignment
or memcpy in an entry block.