snc: allow var_expr for event flags in sync, syncq, and pv init
It was already allowed to to declare event flags in aggregates (structs,
arrays, etc) and to reference them in action code. Now we can also reference
them in sync and syncq clauses, as well as pv initialisers, with the usual
restriction that only array subscripts and struct member selection may occur
in the expression.