snc: use pv_type_check for stricter checking when a pv type is expected
This is necessary because in the generated C code all channel objects have
the same type, so the C compiler cannot do the type checking for us. A
subtlety here is that when we wrap the expected type with pv on the fly for
the argument of seq_pvValue, then we have to loosen the type checking
because what we are really after is the value in this case, not the channel
object.