back to generating a channel table
It turned out that creating a function call for each channel (and each event
flag and each monitor for each channel) can be costly with regard to compile
time. Generating nested for loops for channel arrays would be possible but
is much more complicated than the simple table approach.

Support for state set specific monitors was previously implemented by
generating a call for each channel and state set. We now generate monitor
masks, where a bit represents a state set that has a monitor on the channel.
Only masks that are actually used for some channel are generated, and only
once per set of state sets. This accounts for the bulk of the added code in
gen_tables.c.

The functions seq_pvCreate, seq_efCreate, and seq_pvAddMonitor are no longer
public but private to the runtime library.